automations update

+++ app/main/routes.py
@@ -11,13 +11,16 @@ def index():
 
 @bp.route('/automations', methods=['GET', 'POST'])
 def automations():
+    if Fan.query.count() == 0 or TempSensor.query.count() == 0:
+      flash(f'Must have at least one fan and one Temp Sensor for automations')
+      redirect(url_for('main.index'))
     if Automation.query.count() == 0:
-      return redirect(url_for('.newautomation'))
+      return redirect(url_for('.edit_automation')+'?name=_new_')
     form = AutomationForm(request.form)
     if form.validate_on_submit():
       if auto := Automation.query.filter_by(name=form.name.data).first():
         if form.edit.data == True:
-          return redirect(url_for('.editautomation')+'?name='+auto.name)
+          return redirect(url_for('.edit_automation')+'?name='+auto.name)
         elif form.delete.data == True:
           db.session.delete(auto)
           flash(f'DELETED Automation: {form.name.data}')
@@ -28,40 +31,27 @@ def automations():
     forms = [AutomationForm(formdata=None, obj=auto) for auto in Automation.query.all()]
     return render_template('automations.html', title='Automations', forms=forms)
 
-@bp.route('/newautomation', methods=['GET', 'POST'])
-def newautomation():
-    if Fan.query.count() == 0:
-      flash(f'Need to create at least one fan before creating automations')
+@bp.route('/edit_automation', methods=['GET', 'POST'])
+def edit_automation():
+    if Fan.query.count() == 0 or TempSensor.query.count() == 0:
+      flash(f'Must have at least one fan and one Temp Sensor for automations')
       redirect(url_for('main.index'))
-    if TempSensor.query.count() == 0:
-      flash(f'Need to create at least one Temp Sensor before creating automations')
-      redirect(url_for('main.index'))
-    form = EditAutomationForm(request.form, disp_title='New Automation')
+    if (name := request.args.get('name')) == '_new_':
+      auto = Automation()
+      prefixes = ('New', 'Created new')
+    else:
+      auto = Automation.query.filter_by(name=name).first()
+      if not auto:
+        flash('Error: automation {name} was not found')
+        return redirect(url_for('.automations'))
+      prefixes = ('Edit', 'Edited')
+    form = EditAutomationForm(obj=auto, disp_title=prefixes[0]+' automation')
     if form.validate_on_submit():
-      # if auto := Automation.query.filter_by(name=form.name.data).first():
-      auto = Automation(form)
+      if form.cancel.data: 
+        return redirect(url_for('.automations'))
+      auto.copy_from_form(form)
       db.session.add(auto)
       db.session.commit()
-      flash(f'Created new automation {auto.name}')
-      return redirect(url_for('main.automations'))
-    elif request.method == 'GET':
-       form.sensor_name.choices = [sensor.name for sensor in TempSensor.query.all()]
-       form.fan_name.choices = [fan.name for fan in Fan.query.all()]
-    return render_template('edit_automation.html', title='New Automation', form=form)
-
-@bp.route('/editautomation', methods=['GET', 'POST'])
-def editautomation():
-    if not (auto := Automation.query.filter_by(name=request.args.get('name')).first()):
-      flash(f'Fan {request.args.get("name")} not found')
-      return redirect(url_for('fans.fans_index'))
-      #todo: handle errors better
-    form = EditAutomationForm(obj=auto, disp_title='Edit Automation')
-    if form.validate_on_submit():
-      auto.copy_from_form(form)
-      db.session.commit()
-      flash(f'Edited automation {auto.name}')
-      return redirect(url_for('main.automations'))
-    elif request.method == 'GET':
-       form.sensor_name.choices = [sensor.name for sensor in TempSensor.query.all()]
-       form.fan_name.choices = [fan.name for fan in Fan.query.all()]
-    return render_template('edit_automation.html', title='Edit Automation', form=form)
+      flash(f'{prefixes[1]} automation named {auto.name}')
+      return redirect(url_for('.automations'))
+    return render_template('edit_automation.html', title=f'{prefixes[0]} automation', form=form)
